// src/app/languages/interface/auth/changepassword.interface.ts
export interface IChangePassword {
    oldPassword: string;
    enterOldPassword: string;
    newPassword: string;
    enterNewPassword: string;
    confirmPassword: string;
    enterConfirmPassword: string;
    submit: string;
    loading: string;
    back: string;
    oldpassisreq: string;
    newpassisreq: string;
    confirmpassisreq: string
}